// gui/src/schema/ContextItemId.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type ContextItemId = ContextItemId1;
export type ProviderTitle = string;
export type ItemId = string;

/**
 * A ContextItemId is a unique identifier for a ContextItem.
 */
export interface ContextItemId1 {
  provider_title: ProviderTitle;
  item_id: ItemId;
  [k: string]: unknown;
}